FIX: initial size of component editor is 0,0
A traitsui item with editor=ComponentEditor is
displayed in Qt with initial size equals to 0, 0.
For example, a dialog containing a Chaco plot will
not even display the frame of the window, unless
the width and height of the view is set.

The commit contains test and fix.
Fixes Issues 45 and 46 in Chaco.