Modified draw_image to work around memory leak in QPainter.device().
Instead of calling QPainter.device() to check that the painter is active
and to get the device, the kiva.qpainter.GraphicsContext qt_dc member is
checked to see that it is an instance of QPixmap. To check that the
painter is active, QPainter.isActive() is called on
kiva.qpainter.GraphicsContext.gc . This has been tested with
qt_example.py  in the Chaco demos by setting: ETSConfig.toolkit =
"qt4.qpainter" .
Where there was once a memory leak, it now appears that there is none.