Move requirements for EDM and remove installing demo dependencies
Instead of installing the same releases of traits/traitsui/pyface twice,
set the requirements for EDM.

The installation for demo dependencies (see #386) was removed. They
are not used for CI, and are rarely used for development.
Installing demo dependencies brings in the released egg of enable,
causing CI to stop testing local enable source.