FIX : Provide the right path to the version file on windows
On the main branch, the _version.py files dont get generated in enable
and kiva - because of the path provided to write_version_py function.
The changes made here enable the generation of the version file - which
also fixes an issue where documentation build is broken on windows
because of the lack of the version file.

	modified:   setup.py