Backport PRs #888 and #889 to 5.2 maintenance branch (#890)
* Fix svg backend issue with pyparsing version >= 3 (#888)

* FIX : Fix issue with pyparsing version >= 3

fixes #887 

This PR fixes an issue because of an upstream change with
the pyparsing CaselessLiteral class, where the name attribute became
a read-only property in version >= 3. We now use the new public
set_name method instead of directly setting the attribute.

* FIX : Handle pyparsing versions 2 and 3 accordingly

the new set_name API is only available on pyparsing version >= 3 so we
use packaging to handle the behavior according to the version

this commit makes packaging a dependency of the svg backend

and this commit adds a regression test

	modified:   enable/__init__.py
	modified:   enable/savage/svg/pathdata.py
	modified:   enable/savage/svg/tests/test_pathdata.py

* REF : Go with a simpler solution of checking for the method name

instead of manually parsing and checking the pyparsing version and
calling the relevant method depending on the version, we simply check
for the existence of the set_name method now and use it if it exists. If
it doesnt, we fall back to the old setName method

The dependence of packaging has now been removed because we dont need to
parse the version anymore

	modified:   enable/__init__.py
	modified:   enable/savage/svg/pathdata.py

* Test on ubuntu-18.04, not 16.04 (#889)

* FIX : Test on ubuntu-18.04, not 16.04

ubuntu-16.04 is no longer available on GitHub Actions CI

* FIX : Use ubuntu-latest instead of ubuntu-18.04

* FIX : Use ubuntu-18.04 with the right package repository

	modified:   .github/workflows/test-with-edm.yml
	modified:   ci/edmtool.py

* FIX : Install libsdl2-dev using the ubuntu package manager

	modified:   .github/workflows/test-with-edm.yml

* FIX : Use libsdl1.2, not libsdl2

	modified:   .github/workflows/test-with-edm.yml