Ensure toolkit is initialized before use (#921)
* Ensure toolkit is initialized before use

In a number of places the code used ETSConfig.toolkit to determine the backend
at import time.  However it is possible that that value was being checked
before ETSConfig.toolkit was properly set.  This changes the code to get the
appropriate Pyface or TraitsUI toolkit object and gets the name of the toolkit
from that.

This fixes #913.

* Remove a few more uses of ETSConfig

Also fix a couple more uses of ETSConfig.toolkit.

* Fix tests on PySide2

* More fixes for tests.

* Protect against imports from pyface or triatsui in Kiva if not installed.