Refactor color traits (#924)
* Refactor color traits

This combines the two color traits in Enable: ColorTrait (which is mapped)
and RGBAColorTrait (which is not), updating them so that they don't use the
old `Trait` factory and are simple TraitTypes instead.  It combines validation
so that they accept the same wide set of values, and makes use of the code from
Pyface for parsing color values and the standard color names (Pyface's values
were originally taken from Enable).  It also unifies the editors, so that both
traits have the same editors.

* Prevent circular imports in Wx color editor code.

* Apply suggestions from code review

Co-authored-by: Mark Dickinson <mdickinson@enthought.com>

* Various improvements suggested from code review.

* Fix imports.

Co-authored-by: Mark Dickinson <mdickinson@enthought.com>