Upgrade workflow to Ubuntu 20.04 (#982)
This PR updates the bleeding-edge.yml workflow to use Ubuntu 20.04 instead of Ubuntu 18.04; the reason for using 18.04 (that 20.04 has SWIG 4) is no longer relevant now that Enable supports building with SWIG 4.

The real reason I'm making this PR is that I'm hoping it'll magically fix issue #981. It probably won't, but it seems worth a try before diving into more time-consuming debugging.