Fix CI (#1036)
This gets a reasonable baseline CI working for Kiva and Enable.  It covers Python 3.8 and 3.10 with a reasonable set of Kiva backends and ETS toolkits.  It splits out the Linux Qt dependency installation into a separate action as is done in other ETS projects so it can be re-used across workflows.

There are a number of test cases skipped:

- WxPython on MacOS and Linux due to lack of appropriate wheels
- PyQt5 on Python 3.10 due to #1037
- PyQt5 on Windows due to #1038

This also adds a skip to a Cairo backend test case due to #1035 and a couple of drive-by fixes to the Cairo Kiva backend (Cairo was previously not being tested).