Add an action that builds wheels (#1039)
This adds a stand-alone build wheel action, and then adds it to the publish to PyPI action for releases. This builds wheels for Mac, Windows and Linux Intel architecture, as well as Linux ARM.

It doesn't handle Mac ARM or Universal, but it potentially could with some more work and testing.