Map `MODERN` to `"monospace"` and `TELETYPE` to `"typewriter"` (#977)
This changes font handling to:
- map "modern" font family to "monospace" rather than "sans-serif" like other backends
- distinguish between "monospace" (sans-serif fixed pitch fonts) and "typewriter" (serif fixed pitch fonts).

Replaces uses of "modern" with "sans-serif" where appropriate (particularly in defaults) to keep appearances the same.

This is not backwards compatible: users who use "modern" font family will get different fonts, but it is easily fixed by changing the font family.  So this should probably be part of a major release.

Fixes #974.